+++ app/modules/AdminModule/Presenters/HomePresenter.php
@@ -3,7 +3,6 @@
 namespace App\Modules\AdminModule;
 
 use App\Components\Widgets\AboutContainerWidget\AboutContainerWidget;
-use App\Constants\ContainerEnvironments;
 use App\Constants\SessionNames;
 use App\Core\Datetypes\DateTime;
 use App\Core\Http\HttpRequest;
@@ -60,10 +59,6 @@ class HomePresenter extends AAdminPresenter {
         $form->addDateTimeInput('containerDateCreated', 'Date created:')
             ->setDisabled()
             ->setValue($dateCreated);
-
-        $form->addTextInput('containerEnvironment', 'Container environment:')
-            ->setDisabled()
-            ->setValue(ContainerEnvironments::toString($container->getEnvironment()));
 
         return $form;
     }