+++ data/db/migrations/migration_2025-06-27_0018_container_environments_removal.php
@@ -1,0 +1,28 @@
+<?php
+
+namespace App\Data\Db\Migrations;
+
+use App\Core\DB\ABaseMigration;
+use App\Core\DB\Helpers\TableSchema;
+use App\Core\DB\Helpers\TableSeeding;
+
+class migration_2025_06_27_0018_container_environments_removal extends ABaseMigration {
+    public function up(): TableSchema {
+        $schema = $this->getTableSchema();
+
+        $schema->update('containers')
+            ->removeColumn('environment');
+
+        return $schema;
+    }
+
+    public function down(): TableSchema {
+        return $this->getTableSchema();
+    }
+
+    public function seeding(): TableSeeding {
+        return $this->getTableSeeding();
+    }
+}
+
+?>